+++ db/migrate/20210714132606_remove_roda_identifier_fragment_from_activity.rb
@@ -1,0 +1,5 @@
+class RemoveRodaIdentifierFragmentFromActivity < ActiveRecord::Migration[6.1]
+  def change
+    remove_column :activities, :roda_identifier_fragment, :string
+  end
+end