+++ spec/lib/tasks/invalid_activities_spec.rb
@@ -14,13 +14,13 @@ RSpec.describe "rake invalid_activities" do
     expect(invalid_activities_from_csv.count).to eql(2)
 
     aggregate_failures do
-      first_invalid_activity_row = invalid_activities_from_csv.detect { |row| row.include?(first_invalid_activity.roda_identifier_compound) }
+      first_invalid_activity_row = invalid_activities_from_csv.detect { |row| row.include?(first_invalid_activity.roda_identifier) }
       expect(first_invalid_activity_row).to include(first_invalid_activity.title)
       expect(first_invalid_activity_row).to include(first_invalid_activity.organisation.name)
       expect(first_invalid_activity_row).to include(first_invalid_activity.level)
       expect(first_invalid_activity_row).to include(Rails.application.routes.url_helpers.organisation_activity_details_url(first_invalid_activity.organisation, first_invalid_activity.id, host: ENV["DOMAIN"]))
 
-      second_invalid_activity_row = invalid_activities_from_csv.detect { |row| row.include?(second_invalid_activity.roda_identifier_compound) }
+      second_invalid_activity_row = invalid_activities_from_csv.detect { |row| row.include?(second_invalid_activity.roda_identifier) }
       expect(second_invalid_activity_row).to include(second_invalid_activity.title)
       expect(second_invalid_activity_row).to include(second_invalid_activity.organisation.name)
       expect(second_invalid_activity_row).to include(second_invalid_activity.level)