+++ spec/models/activity_spec.rb
@@ -188,162 +188,6 @@ RSpec.describe Activity, type: :model do
       end
     end
 
-    describe "#roda_identifier_fragment" do
-      context "for a fund" do
-        let(:fund) { create(:fund_activity) }
-
-        it "is valid if blank" do
-          fund.roda_identifier_fragment = ""
-          expect(fund).to be_valid(:roda_identifier_step)
-        end
-
-        it "is valid when containing certain punctuation" do
-          fund.roda_identifier_fragment = "A-_/\\"
-          expect(fund).to be_valid(:roda_identifier_step)
-        end
-
-        it "is not valid when containing other punctuation" do
-          fund.roda_identifier_fragment = "A!A"
-          expect(fund).not_to be_valid(:roda_identifier_step)
-        end
-
-        it "is valid up to 5 chars" do
-          fund.roda_identifier_fragment = "A" * 5
-          expect(fund).to be_valid(:roda_identifier_step)
-        end
-
-        it "is not valid above 6 chars" do
-          fund.roda_identifier_fragment = "A" * 7
-          expect(fund).not_to be_valid(:roda_identifier_step)
-        end
-      end
-
-      context "for a programme" do
-        let(:fund) { create(:fund_activity) }
-        let(:programme) { create(:programme_activity, parent: fund) }
-
-        before do
-          fund.update!(roda_identifier_fragment: "A" * 5)
-        end
-
-        it "is valid if blank" do
-          programme.roda_identifier_fragment = ""
-          expect(programme).to be_valid(:roda_identifier_step)
-        end
-
-        it "is valid when containing certain punctuation" do
-          programme.roda_identifier_fragment = "B-B_B/B\\123"
-          expect(programme).to be_valid(:roda_identifier_step)
-        end
-
-        it "is not valid when containing other punctuation" do
-          programme.roda_identifier_fragment = "B!B"
-          expect(programme).not_to be_valid(:roda_identifier_step)
-        end
-
-        it "is valid if 'A-B' is up to 18 chars" do
-          programme.roda_identifier_fragment = "B" * 12
-          expect(programme).to be_valid(:roda_identifier_step)
-        end
-
-        it "is not valid if 'A-B' exceeds 18 chars" do
-          programme.roda_identifier_fragment = "B" * 13
-          expect(programme).not_to be_valid(:roda_identifier_step)
-        end
-
-        context "when there are other activities with similar identifiers" do
-          let!(:fund_a) { create(:fund_activity, roda_identifier_fragment: "AAA") }
-          let!(:programme_a) { create(:programme_activity, parent: fund_a, roda_identifier_fragment: "B-CCC") }
-
-          let!(:fund_b) { create(:fund_activity, roda_identifier_fragment: "AAA-B") }
-
-          it "is invalid if it has the same fragment identifier as a sibling activity" do
-            programme = build(:programme_activity, parent: fund_a, roda_identifier_fragment: "B-CCC")
-            programme.cache_roda_identifier!
-
-            expect(programme).not_to be_valid
-          end
-
-          it "is invalid if it has the same compound identifier as a cousin activity" do
-            programme = build(:programme_activity, parent: fund_b, roda_identifier_fragment: "CCC")
-            programme.cache_roda_identifier!
-
-            expect(programme).not_to be_valid
-          end
-
-          it "is valid if its identifier is different to all others" do
-            programme = build(:programme_activity, parent: fund_b, roda_identifier_fragment: "VALID")
-            programme.cache_roda_identifier!
-
-            expect(programme).to be_valid
-          end
-        end
-      end
-
-      context "for a project" do
-        let(:project) { create(:project_activity) }
-
-        it "is valid if blank" do
-          project.roda_identifier_fragment = ""
-          expect(project).to be_valid(:roda_identifier_step)
-        end
-
-        it "is valid when containing certain punctuation" do
-          project.roda_identifier_fragment = "C-C_C/C\\123"
-          expect(project).to be_valid(:roda_identifier_step)
-        end
-
-        it "is not valid when containing other punctuation" do
-          project.roda_identifier_fragment = "C!C"
-          expect(project).not_to be_valid(:roda_identifier_step)
-        end
-
-        it "is valid up to 20 chars" do
-          project.roda_identifier_fragment = "C" * 20
-          expect(project).to be_valid(:roda_identifier_step)
-        end
-
-        it "is not valid above 20 chars" do
-          project.roda_identifier_fragment = "C" * 21
-          expect(project).not_to be_valid(:roda_identifier_step)
-        end
-      end
-
-      context "for a third-party project" do
-        let(:project) { create(:project_activity) }
-        let(:third_party_project) { create(:third_party_project_activity, parent: project) }
-
-        before do
-          project.update!(roda_identifier_fragment: "C" * 10)
-        end
-
-        it "is valid if blank" do
-          third_party_project.roda_identifier_fragment = ""
-          expect(third_party_project).to be_valid(:roda_identifier_step)
-        end
-
-        it "is valid when containing certain punctuation" do
-          third_party_project.roda_identifier_fragment = "D-D_D/D\\123"
-          expect(third_party_project).to be_valid(:roda_identifier_step)
-        end
-
-        it "is not valid when containing other punctuation" do
-          third_party_project.roda_identifier_fragment = "D!D"
-          expect(third_party_project).not_to be_valid(:roda_identifier_step)
-        end
-
-        it "is valid if 'CD' is up to 18 chars" do
-          third_party_project.roda_identifier_fragment = "D" * 11
-          expect(third_party_project).to be_valid(:roda_identifier_step)
-        end
-
-        it "is not valid if 'CD' exceeds 18 chars" do
-          third_party_project.roda_identifier_fragment = "D" * 12
-          expect(third_party_project).not_to be_valid(:roda_identifier_step)
-        end
-      end
-    end
-
     context "when title is blank" do
       subject(:activity) { build(:project_activity, title: nil) }
       it "should not be valid" do
@@ -1147,74 +991,6 @@ RSpec.describe Activity, type: :model do
     end
   end
 
-  describe "#transparency_identifier" do
-    context "when the activity is a fund" do
-      it "returns a composite identifier formed with the service owner" do
-        fund = create(:fund_activity, roda_identifier_fragment: "GCRF1")
-
-        expected_identifier = [
-          Organisation::SERVICE_OWNER_IATI_REFERENCE,
-          fund.roda_identifier_fragment,
-        ].join("-")
-
-        expect(fund.transparency_identifier).to eql(expected_identifier)
-      end
-    end
-
-    context "when the activity is a programme" do
-      it "returns an identifier with the service owner, fund and programme" do
-        programme = create(:programme_activity)
-        fund = programme.parent
-
-        expected_identifier = [
-          Organisation::SERVICE_OWNER_IATI_REFERENCE,
-          fund.roda_identifier_fragment,
-          programme.roda_identifier_fragment,
-        ].join("-")
-
-        expect(programme.transparency_identifier)
-          .to eql(expected_identifier)
-      end
-    end
-
-    context "when the activity is a project" do
-      it "returns an identifier with the service owner, fund, programme and project" do
-        project = create(:project_activity)
-        programme = project.parent
-        fund = programme.parent
-
-        expected_identifier = [
-          Organisation::SERVICE_OWNER_IATI_REFERENCE,
-          fund.roda_identifier_fragment,
-          programme.roda_identifier_fragment,
-          project.roda_identifier_fragment,
-        ].join("-")
-
-        expect(project.transparency_identifier)
-          .to eql(expected_identifier)
-      end
-    end
-
-    context "when the activity is a third-party project" do
-      it "returns an identifier with the service owner, fund, programme, project and third-party project" do
-        third_party_project = create(:third_party_project_activity)
-        project = third_party_project.parent
-        programme = project.parent
-        fund = programme.parent
-
-        project_identifier = [
-          Organisation::SERVICE_OWNER_IATI_REFERENCE,
-          fund.roda_identifier_fragment,
-          programme.roda_identifier_fragment,
-          project.roda_identifier_fragment,
-        ].join("-")
-
-        expect(third_party_project.transparency_identifier)
-          .to eql("#{project_identifier}#{third_party_project.roda_identifier_fragment}")
-      end
-    end
-  end
-
   describe "#iati_identifier" do
     it "returns the previous_identifier if it exists" do
       activity = create(:project_activity, previous_identifier: "previous-id", transparency_identifier: "transparency-id")
@@ -1224,107 +1000,6 @@ RSpec.describe Activity, type: :model do
     it "returns the transparency_identifier if previous_identifier is not set" do
       activity = create(:project_activity, previous_identifier: nil, transparency_identifier: "transparency-id")
       expect(activity.iati_identifier).to eq("transparency-id")
-    end
-  end
-
-  describe "#can_set_roda_identifier?" do
-    let!(:fund) { create(:fund_activity, roda_identifier_fragment: "Lvl/A") }
-    let!(:programme) { create(:programme_activity, parent: fund, roda_identifier_fragment: "Level/B") }
-    let!(:project) { create(:project_activity, parent: programme, roda_identifier_fragment: nil) }
-
-    context "for a top-level (fund) activity" do
-      it "is true when the activity does not have a RODA identifier" do
-        fund.roda_identifier_fragment = nil
-        expect(fund.can_set_roda_identifier?).to be(true)
-      end
-
-      it "is false when the activity already has a RODA identifier" do
-        expect(fund.can_set_roda_identifier?).to be(false)
-      end
-    end
-
-    it "is true when all parent identifiers are present" do
-      expect(project.can_set_roda_identifier?).to be(true)
-    end
-
-    it "is false if the activity has a RODA identifier" do
-      project.update!(roda_identifier_fragment: "Level/C")
-      expect(project.can_set_roda_identifier?).to be(false)
-    end
-
-    it "is false if the parent identifier is missing" do
-      programme.update!(roda_identifier_fragment: nil)
-      expect(project.can_set_roda_identifier?).to be(false)
-    end
-
-    it "is false if the grandparent identifier is missing" do
-      fund.update!(roda_identifier_fragment: nil)
-      expect(project.can_set_roda_identifier?).to be(false)
-    end
-  end
-
-  describe "#cache_roda_identifier!" do
-    let!(:fund) { create(:fund_activity, roda_identifier_fragment: "Lvl/A") }
-    let!(:programme) { create(:programme_activity, parent: fund, roda_identifier_fragment: "Level/B") }
-    let!(:project) { create(:project_activity, parent: programme, roda_identifier_fragment: "Level/C") }
-    let!(:third_party_project) { create(:third_party_project_activity, parent: project, roda_identifier_fragment: "Level/D") }
-
-    before do
-      project.write_attribute(:roda_identifier_compound, nil)
-      third_party_project.write_attribute(:roda_identifier_compound, nil)
-    end
-
-    it "raises an exception if roda_identifier_compound is overwritten" do
-      expect { fund.cache_roda_identifier! }.to raise_error(TypeError, "Activity #{fund.id} already has a compound RODA identifier")
-    end
-
-    it "caches the compound RODA identifier on a project" do
-      project.cache_roda_identifier!
-      expect(project.roda_identifier_compound).to eq("Lvl/A-Level/B-Level/C")
-    end
-
-    it "caches the transparency identifier on a project" do
-      project.cache_roda_identifier!
-      expect(project.transparency_identifier).to eq("GB-GOV-13-Lvl-A-Level-B-Level-C")
-    end
-
-    it "caches the compound RODA identifier on a third-party project" do
-      third_party_project.cache_roda_identifier!
-      expect(third_party_project.roda_identifier_compound).to eq("Lvl/A-Level/B-Level/CLevel/D")
-    end
-
-    context "when the activity does not have a RODA identifier fragment" do
-      before { project.update!(roda_identifier_fragment: nil) }
-
-      it "raises an exception" do
-        expect { project.cache_roda_identifier! }.to raise_error(TypeError)
-      end
-    end
-
-    context "when the activity's parent does not have a RODA identifier fragment" do
-      before { programme.update!(roda_identifier_fragment: nil) }
-
-      it "raises an exception" do
-        expect { project.cache_roda_identifier! }.to raise_error(TypeError)
-      end
-    end
-
-    context "when the activity's grandparent does not have a RODA identifier fragment" do
-      before { fund.update!(roda_identifier_fragment: nil) }
-
-      it "raises an exception" do
-        expect { project.cache_roda_identifier! }.to raise_error(TypeError)
-      end
-    end
-  end
-
-  describe "#roda_identifier" do
-    let(:fund) { create(:fund_activity, roda_identifier_fragment: "GCRF") }
-    let(:programme) { create(:programme_activity, parent: fund, roda_identifier_fragment: "RSDel") }
-    let(:project) { create(:project_activity, parent: programme, roda_identifier_fragment: "DEL_Misc") }
-
-    it "returns the compound RODA identifier" do
-      expect(project.roda_identifier).to eq("GCRF-RSDel-DEL_Misc")
     end
   end
 
@@ -1500,21 +1175,21 @@ RSpec.describe Activity, type: :model do
 
   describe ".hierarchically_grouped_projects" do
     before do
-      first_project = create(:project_activity, roda_identifier_fragment: "zzxx")
-      create(:third_party_project_activity, roda_identifier_fragment: "ww", parent: first_project)
+      first_project = create(:project_activity, roda_identifier: "zzxx")
+      create(:third_party_project_activity, roda_identifier: "ww", parent: first_project)
 
-      _second_project = create(:project_activity, roda_identifier_fragment: "mmnn")
+      _second_project = create(:project_activity, roda_identifier: "mmnn")
 
-      third_project = create(:project_activity, roda_identifier_fragment: "aabb")
+      third_project = create(:project_activity, roda_identifier: "aabb")
       (1..3).each do |i|
-        create(:third_party_project_activity, roda_identifier_fragment: "cc#{3 - i}", parent: third_project)
+        create(:third_party_project_activity, roda_identifier: "cc#{3 - i}", parent: third_project)
       end
     end
 
     it "returns projects followed by their third-party project children" do
       result = Activity.all.hierarchically_grouped_projects
 
-      expect(result.map(&:roda_identifier_fragment)).to eq(["aabb", "cc0", "cc1", "cc2", "mmnn", "zzxx", "ww"])
+      expect(result.map(&:roda_identifier)).to eq(["aabb", "cc0", "cc1", "cc2", "mmnn", "zzxx", "ww"])
     end
   end