+++ spec/services/activities/import_from_csv_spec.rb
@@ -15,9 +15,8 @@ RSpec.describe Activities::ImportFromCsv do
   end
   let(:existing_activity_attributes) do
     {
-      "RODA ID" => existing_activity.roda_identifier_compound,
+      "RODA ID" => existing_activity.roda_identifier,
       "Transparency identifier" => "13232332323",
-      "RODA ID Fragment" => "",
       "Parent RODA ID" => "",
       "Title" => "Here is a title",
       "Description" => "Some description goes here...",
@@ -69,8 +68,7 @@ RSpec.describe Activities::ImportFromCsv do
   let(:new_activity_attributes) do
     existing_activity_attributes.merge({
       "RODA ID" => "",
-      "RODA ID Fragment" => "234566",
-      "Parent RODA ID" => parent_activity.roda_identifier_compound,
+      "Parent RODA ID" => parent_activity.roda_identifier,
       "Transparency identifier" => "23232332323",
     })
   end
@@ -101,24 +99,8 @@ RSpec.describe Activities::ImportFromCsv do
       expect(subject.errors.first.message).to eq(I18n.t("importer.errors.activity.not_found"))
     end
 
-    it "has an error when the ID present, but there is a fragment present" do
-      existing_activity_attributes["RODA ID Fragment"] = "13344"
-
-      expect { subject.import([existing_activity_attributes]) }.to_not change { existing_activity }
-
-      expect(subject.created.count).to eq(0)
-      expect(subject.updated.count).to eq(0)
-
-      expect(subject.errors.count).to eq(1)
-      expect(subject.errors.first.csv_row).to eq(2)
-      expect(subject.errors.first.csv_column).to eq("RODA ID Fragment")
-      expect(subject.errors.first.column).to eq(:roda_identifier_fragment)
-      expect(subject.errors.first.value).to eq("13344")
-      expect(subject.errors.first.message).to eq(I18n.t("importer.errors.activity.cannot_update.fragment_present"))
-    end
-
     it "has an error when the ID present, but there is a parent present" do
-      existing_activity_attributes["Parent RODA ID"] = parent_activity.roda_identifier_fragment
+      existing_activity_attributes["Parent RODA ID"] = parent_activity.roda_identifier
 
       expect { subject.import([existing_activity_attributes]) }.to_not change { existing_activity }
 
@@ -129,7 +111,7 @@ RSpec.describe Activities::ImportFromCsv do
       expect(subject.errors.first.csv_row).to eq(2)
       expect(subject.errors.first.csv_column).to eq("Parent RODA ID")
       expect(subject.errors.first.column).to eq(:parent_id)
-      expect(subject.errors.first.value).to eq(parent_activity.roda_identifier_fragment)
+      expect(subject.errors.first.value).to eq(parent_activity.roda_identifier)
       expect(subject.errors.first.message).to eq(I18n.t("importer.errors.activity.cannot_update.parent_present"))
     end
 
@@ -232,7 +214,7 @@ RSpec.describe Activities::ImportFromCsv do
     it "has an error and does not update any other activities if a region does not exist" do
       activity_2 = create(:project_activity)
       invalid_activity_attributes = existing_activity_attributes.merge({
-        "RODA ID" => activity_2.roda_identifier_compound,
+        "RODA ID" => activity_2.roda_identifier,
         "Recipient Region" => "111111",
       })
 
@@ -275,7 +257,7 @@ RSpec.describe Activities::ImportFromCsv do
 
       let(:attributes) do
         attributes = existing_activity_attributes.map { |k, _v| [k, ""] }.to_h
-        attributes["RODA ID"] = existing_activity.roda_identifier_compound
+        attributes["RODA ID"] = existing_activity.roda_identifier
         attributes
       end
       let(:changed_attributes) do
@@ -417,19 +399,12 @@ RSpec.describe Activities::ImportFromCsv do
 
       new_activity = Activity.order(:created_at).last
 
-      expected_roda_identifier_compound = [
-        parent_activity.roda_identifier_compound,
-        new_activity_attributes["RODA ID Fragment"],
-      ].join("-")
-
       expect(new_activity.parent).to eq(parent_activity)
       expect(new_activity.source_fund_code).to eq(1)
       expect(new_activity.level).to eq("project")
-      expect(new_activity.roda_identifier_compound).to eq(expected_roda_identifier_compound)
       expect(new_activity.transparency_identifier).to eq(new_activity_attributes["Transparency identifier"])
       expect(new_activity.title).to eq(new_activity_attributes["Title"])
       expect(new_activity.description).to eq(new_activity_attributes["Description"])
-      expect(new_activity.roda_identifier_fragment).to eq(new_activity_attributes["RODA ID Fragment"])
       expect(new_activity.recipient_region).to eq(new_activity_attributes["Recipient Region"])
       expect(new_activity.recipient_country).to eq(new_activity_attributes["Recipient Country"])
       expect(new_activity.intended_beneficiaries).to eq(["KH", "KP", "ID"])
@@ -728,7 +703,7 @@ RSpec.describe Activities::ImportFromCsv do
 
     context "when the activity is a project" do
       it "has an error if the 'Channel of delivery code' is empty" do
-        new_activity_attributes["Parent RODA ID"] = parent_activity.roda_identifier_compound
+        new_activity_attributes["Parent RODA ID"] = parent_activity.roda_identifier
         new_activity_attributes["Channel of delivery code"] = ""
 
         expect { subject.import([new_activity_attributes]) }.to_not change { Activity.count }
@@ -832,22 +807,6 @@ RSpec.describe Activities::ImportFromCsv do
       expect(subject.errors.first.column).to eq(:parent_id)
       expect(subject.errors.first.value).to eq("111111")
       expect(subject.errors.first.message).to eq(I18n.t("importer.errors.activity.parent_not_found"))
-    end
-
-    it "has an error if the identifier is invalid" do
-      new_activity_attributes["RODA ID Fragment"] = "%^$!234566"
-
-      expect { subject.import([new_activity_attributes]) }.to_not change { Activity.count }
-
-      expect(subject.created.count).to eq(0)
-      expect(subject.updated.count).to eq(0)
-
-      expect(subject.errors.count).to eq(1)
-      expect(subject.errors.first.csv_row).to eq(2)
-      expect(subject.errors.first.csv_column).to eq("RODA ID Fragment")
-      expect(subject.errors.first.column).to eq(:roda_identifier_fragment)
-      expect(subject.errors.first.value).to eq("%^$!234566")
-      expect(subject.errors.first.message).to eq(I18n.t("activerecord.errors.models.activity.attributes.roda_identifier_fragment.invalid_characters"))
     end
 
     context "implementing organisation" do